requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pymysql
sqlalchemy
requests
ddgs
dbutils